+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG for PHP CS Fixer: custom fixers
 
+## [Unreleased]
+- Add NoUselessCommentFixer
+- Deprecate NoUselessClassCommentFixer
+- Deprecate NoUselessConstructorCommentFixer
+
 ## v1.11.0 - *2018-10-14*
 - Add PhpdocParamOrderFixer
 - Add InternalClassCasingFixer